requirements.txt
Django==2.0.3
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.6.0
python-decouple==3
Pillow==3.1.2
djangorestframework==3.8.1
djangorestframework-jwt==1.11.0
whitenoise==3.3.1
psycopg2==2.7.4
psycopg2-binary==2.7.4

